requirements.txt
invoke==2.2.0
loguru==0.7.3
Markdown==3.7
pandas==2.2.3
Pillow==11.1.0
pydantic==2.10.5
timm==1.0.13
torch==2.5.1
torchvision==0.20.1
typer==0.15.1
pandas
scikit-learn
dvc
